// src/app/models/Report.ts
export class Report {
    id?: string;
    categoria?: string;
    descripcion?: string;
    fecha?: Date;
    fechaString?: string;
    idUsuario?: string;
    ubicacion?: Location;
    urlImagen?: string;
    nombre?: string;
    hora?: string;
    horario?: string;
    lat?: string;
    long?: string;
}